requirements.txt
google-generativeai
groq
langchain
langchain-community
langchain-core
langchain-groq
langchain-openai
opencv-python
pillow
streamlit 